refactor: 기존에 도메인겸 dto로 사용되고 있던 product를 dto 패키지로 이동
