+++ src/main/java/woowacourse/shoppingcart/domain/OrderDetail.java
@@ -1,5 +1,7 @@
 package woowacourse.shoppingcart.domain;
 
+import woowacourse.shoppingcart.dto.Product;
+
 public class OrderDetail {
     private Long productId;
     private int quantity;
