+++ src/main/java/woowacourse/shoppingcart/dto/OrderRequest.java
@@ -4,8 +4,10 @@ import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
 
 public class OrderRequest {
+
     @NotNull
     private final Long cartId;
+
     @Min(0)
     private final int quantity;
