+++ src/main/java/woowacourse/shoppingcart/dto/Product.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.dto;
 
 public class Product {
     private Long id;
