+++ src/main/java/woowacourse/shoppingcart/ui/CartItemController.java
@@ -5,7 +5,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.domain.Cart;
-import woowacourse.shoppingcart.domain.Product;
+import woowacourse.shoppingcart.dto.Product;
 import woowacourse.shoppingcart.dto.Request;
 import woowacourse.shoppingcart.application.CartService;
